requirements/test.txt
-r base.txt

# Multi-environment testing
tox ==2.9.1

# Wonderful py.test for testing
pytest >=3.6.0
pytest-runner >=2.11.1
pytest-cache >=1.0
pytest-cov >=2.4.0
pytest-mock >=2.0.0

# Faking of file systems
pyfakefs >=3.7.1

# Helper for model creation
factory-boy ==2.12.0

# Easier testing of requests
requests-mock >=1.7.0

# Coverage report
coverage==5.3.1
codacy-coverage >=1.3.6

# Flake8 for linting
flake8 >=3.5.0

# Building documentation is part of the tests
sphinx >=1.7.0
sphinx-rtd-theme >=0.4.3
sphinx-argparse >=0.2.5
m2r

# Type checking.
mypy
